update: add Exodus wallet
Exodus is a non-custodial (or self-custodial) mobile wallet, meaning users retain full control of their private keys and funds, and Exodus does not have access to them. It is not a custodial wallet, where the provider holds the user's private keys. Exodus is available on both desktop and mobile devices, allowing users to manage their crypto assets across platforms.